+++ docs/CHANGELOG.md
@@ -15,6 +15,7 @@ Version 1.2.0 (Willow): Not yet released
 - Add `-afs FILENAME` command line argument to specify a settings file when launching
 - Add `Suppress autoswitch` control bind (if held while picking up a weapon, autoswitch is suppressed)
 - Add `SuppressAutoswitchBindAlias` setting to `alpine_settings.ini`
+- Add `$Flag Captures While Stolen` dedicated server config option
 
 ### Bug fixes
 [@GooberRF](https://github.com/GooberRF)